+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added support for KTX2 texture size reporting in the Asset Size glTF sidebar panel.
 * Added JSON schema for KHR_texture_basisu.
 * Fixed an issue with URI-encoded names. [#200](https://github.com/AnalyticalGraphicsInc/gltf-vscode/issues/200)
+* Fixed an issue with image previews. [#159](https://github.com/AnalyticalGraphicsInc/gltf-vscode/issues/159)
 * Update ThreeJS to r124.
 
 ### 2.3.2 - 2020-12-02